refactor: split NetworkProvider, update gas fees on block change
