+++ apps/protocol/src/app/pages/Save/v2/SaveDeposit.tsx
@@ -3,7 +3,8 @@ import { useMemo, useState } from 'react'
 import { BoostedVault__factory, ISavingsContractV3__factory, SaveWrapper__factory } from '@apps/artifacts/typechain'
 import { AssetExchange, SendButton, TransactionInfo } from '@apps/base/components/forms'
 import { useSigner } from '@apps/base/context/account'
-import { useNetworkAddresses, useNetworkPrices } from '@apps/base/context/network'
+import { useNetworkAddresses } from '@apps/base/context/network'
+import { useNetworkPrices } from '@apps/base/context/prices'
 import { useNativeToken, useTokenSubscription } from '@apps/base/context/tokens'
 import { usePropose } from '@apps/base/context/transactions'
 import { BigDecimal } from '@apps/bigdecimal'
