+++ apps/protocol/src/app/pages/Save/v2/useSaveOutput.ts
@@ -2,7 +2,8 @@ import { useEffect } from 'react'
 
 import { FeederPool__factory, IUniswapV2Router02__factory, Masset__factory, SaveWrapper__factory } from '@apps/artifacts/typechain'
 import { useSigner } from '@apps/base/context/account'
-import { useNetworkAddresses, useNetworkPrices } from '@apps/base/context/network'
+import { useNetworkAddresses } from '@apps/base/context/network'
+import { useNetworkPrices } from '@apps/base/context/prices'
 import { useTokenSubscription } from '@apps/base/context/tokens'
 import { BigDecimal } from '@apps/bigdecimal'
 import { sanitizeMassetError } from '@apps/formatters'
