+++ libs/base/src/lib/context/index.tsx
@@ -6,19 +6,20 @@ import { BlockProvider } from './BlockProvider'
 import { ENSProvider } from './ENSProvider'
 import { NetworkProvider } from './NetworkProvider'
 import { NotificationsProvider } from './NotificationsProvider'
-import { PricesProvider } from './PricesProvider'
+import { NetworkPricesProvider, PricesProvider } from './PricesProvider'
 import { TokensProvider } from './TokensProvider'
 import { TransactionsProvider } from './TransactionsProvider'
 import { WagmiProvider } from './WagmiProvider'
 
 const Providers = composedComponent(
   NetworkProvider,
+  WagmiProvider,
   NotificationsProvider,
   ApolloProvider,
-  WagmiProvider,
   BlockProvider,
   TransactionsProvider,
   TokensProvider,
+  NetworkPricesProvider,
   PricesProvider,
   ENSProvider,
   BannerProvider,
